Fix Clearing Flags
Fixed Clearing of the Flags:
It was not clearing properly awhile ago with the osThreadFlagsWait, so I had switched to just osThreadFlagsClear. This wouldn't work because the "osFlagsNoClear" option has to be setin the wait function. I added this and then needed the manual flag clearing to occur before this wait function, as clearing them within the functions that update flags will not register before the wait function clears the flags again.

Also changed the reset and start button to rising/falling to trigger correctly, probably need to be done similarly to the other GPIO interrupts. This is not reflected in main because my laptop can't generate code correctly.